Update broker command.

+++ internal/ngsicmd/broker.go
@@ -46,16 +46,27 @@ func brokersList(c *cli.Context) error {
 		return &ngsiCmdError{funcName, 1, err.Error(), err}
 	}
 
+	host := c.String("host")
+
+	if host != "" {
+		info, err := ngsi.BrokerList().BrokerInfo(host)
+		if err != nil {
+			return &ngsiCmdError{funcName, 2, host + " not found", err}
+		}
+		printBrokerInfo(ngsi, info)
+		return nil
+	}
+
 	if c.IsSet("json") || c.Bool("pretty") {
 		lists, err := ngsi.BrokerList().BrokerInfoJSON("")
 		if err != nil {
-			return &ngsiCmdError{funcName, 2, err.Error(), err}
+			return &ngsiCmdError{funcName, 3, err.Error(), err}
 		}
 		if c.Bool("pretty") {
 			newBuf := new(bytes.Buffer)
 			err := ngsi.JSONConverter.Indent(newBuf, []byte(*lists), "", "  ")
 			if err != nil {
-				return &ngsiCmdError{funcName, 3, err.Error(), err}
+				return &ngsiCmdError{funcName, 4, err.Error(), err}
 			}
 			fmt.Fprintln(ngsi.StdWriter, string(newBuf.Bytes()))
 		} else {
@@ -104,11 +115,7 @@ func brokersGet(c *cli.Context) error {
 		if err != nil {
 			return &ngsiCmdError{funcName, 5, host + " not found", err}
 		}
-		b, err := ngsilib.JSONMarshal(info)
-		if err != nil {
-			return &ngsiCmdError{funcName, 6, err.Error(), err}
-		}
-		fmt.Fprintln(ngsi.StdWriter, string(b))
+		printBrokerInfo(ngsi, info)
 	}
 
 	return nil
@@ -123,23 +130,27 @@ func brokersAdd(c *cli.Context) error {
 	}
 
 	host := c.String("host")
+	if host == "" {
+		return &ngsiCmdError{funcName, 2, "Required host not found", nil}
+	}
+
 	if ngsilib.IsNameString(host) == false {
-		return &ngsiCmdError{funcName, 2, "name error " + host, err}
+		return &ngsiCmdError{funcName, 3, "name error " + host, err}
 	}
 	if ngsi.ExistsBrokerHost(host) {
-		return &ngsiCmdError{funcName, 3, host + " already exists", err}
+		return &ngsiCmdError{funcName, 4, host + " already exists", err}
 	}
 
 	if !c.IsSet("brokerHost") {
-		return &ngsiCmdError{funcName, 4, "brokerHost is missing", err}
+		return &ngsiCmdError{funcName, 5, "brokerHost is missing", err}
 	}
 
 	if !c.IsSet("ngsiType") && ngsilib.IsHTTP(c.String("brokerHost")) {
-		return &ngsiCmdError{funcName, 5, "ngsiType is missing", err}
+		return &ngsiCmdError{funcName, 6, "ngsiType is missing", err}
 	}
 
 	if c.IsSet("ngsiType") && !ngsilib.IsHTTP(c.String("brokerHost")) {
-		return &ngsiCmdError{funcName, 6, "can't specify ngsiType", err}
+		return &ngsiCmdError{funcName, 7, "can't specify ngsiType", err}
 	}
 
 	param := make(map[string]string)
@@ -159,7 +170,7 @@ func brokersAdd(c *cli.Context) error {
 
 	err = ngsi.CreateBroker(host, param)
 	if err != nil {
-		return &ngsiCmdError{funcName, 7, err.Error(), err}
+		return &ngsiCmdError{funcName, 8, err.Error(), err}
 	}
 	return nil
 }
@@ -167,14 +178,18 @@ func brokersAdd(c *cli.Context) error {
 func brokersUpdate(c *cli.Context) error {
 	const funcName = "brokersUpdate"
 
-	ngsi, err := initCmd(c, funcName, true)
+	ngsi, err := initCmd(c, funcName, false)
 	if err != nil {
 		return &ngsiCmdError{funcName, 1, err.Error(), err}
 	}
 
 	host := c.String("host")
+	if host == "" {
+		return &ngsiCmdError{funcName, 2, "Required host not found", nil}
+	}
+
 	if !ngsi.ExistsBrokerHost(host) {
-		return &ngsiCmdError{funcName, 2, host + " not found", err}
+		return &ngsiCmdError{funcName, 3, host + " not found", err}
 	}
 
 	param := make(map[string]string)
@@ -183,6 +198,9 @@ func brokersUpdate(c *cli.Context) error {
 		key := args[i]
 		if c.IsSet(key) {
 			value := c.String(key)
+			if key == "service" {
+				value = strings.ToLower(value)
+			}
 			if value != "" {
 				param[key] = value
 			}
@@ -191,7 +209,7 @@ func brokersUpdate(c *cli.Context) error {
 
 	err = ngsi.UpdateBroker(host, param)
 	if err != nil {
-		return &ngsiCmdError{funcName, 3, err.Error(), err}
+		return &ngsiCmdError{funcName, 4, err.Error(), err}
 	}
 	return nil
 }
@@ -199,30 +217,34 @@ func brokersUpdate(c *cli.Context) error {
 func brokersDelete(c *cli.Context) error {
 	const funcName = "brokersDelete"
 
-	ngsi, err := initCmd(c, funcName, true)
+	ngsi, err := initCmd(c, funcName, false)
 	if err != nil {
 		return &ngsiCmdError{funcName, 1, err.Error(), err}
 	}
 
 	host := c.String("host")
+	if host == "" {
+		return &ngsiCmdError{funcName, 2, "Required host not found", nil}
+	}
+
 	if !ngsi.ExistsBrokerHost(host) {
-		return &ngsiCmdError{funcName, 2, host + " not found", err}
+		return &ngsiCmdError{funcName, 3, host + " not found", err}
 	}
 
 	if c.IsSet("items") {
 		items := c.String("items")
 		for _, item := range strings.Split(items, ",") {
 			if err := ngsi.DeleteItem(host, item); err != nil {
-				return &ngsiCmdError{funcName, 3, "delete error - " + item, err}
+				return &ngsiCmdError{funcName, 4, "delete error - " + item, err}
 			}
 		}
 		err = ngsi.UpdateBroker(host, nil)
 		if err != nil {
-			return &ngsiCmdError{funcName, 4, err.Error(), err}
+			return &ngsiCmdError{funcName, 5, err.Error(), err}
 		}
 	} else {
 		if err = ngsi.IsHostReferenced(host); err != nil {
-			return &ngsiCmdError{funcName, 5, host + " is referenced", err}
+			return &ngsiCmdError{funcName, 6, host + " is referenced", err}
 		}
 		if ngsi.PreviousArgs.Host == host {
 			ngsi.PreviousArgs.Host = ""
@@ -232,4 +254,63 @@ func brokersDelete(c *cli.Context) error {
 		_ = ngsi.DeleteBroker(host)
 	}
 	return nil
+}
+
+func printBrokerInfo(ngsi *ngsilib.NGSI, info *ngsilib.Broker) {
+
+	fmt.Fprintln(ngsi.StdWriter, "brokerHost "+info.BrokerHost)
+	if info.NgsiType != "" {
+		fmt.Fprintln(ngsi.StdWriter, "ngsiType "+info.NgsiType)
+	}
+	if strings.ToLower(info.NgsiType) == "v2" {
+		if info.Tenant != "" {
+			fmt.Fprintln(ngsi.StdWriter, "FIWARE-Serivce "+info.Tenant)
+		}
+		if info.Scope != "" {
+			fmt.Fprintln(ngsi.StdWriter, "FIWARE-SerivcePath "+info.Scope)
+		}
+	} else {
+		if info.Tenant != "" {
+			fmt.Fprintln(ngsi.StdWriter, "Tenant "+info.Tenant)
+		}
+		if info.Scope != "" {
+			fmt.Fprintln(ngsi.StdWriter, "Scope "+info.Scope)
+		}
+	}
+	if info.Context != "" {
+		fmt.Fprintln(ngsi.StdWriter, "Context "+info.Context)
+	}
+	if info.SafeString != "" {
+		fmt.Fprintln(ngsi.StdWriter, "SafeString "+info.SafeString)
+	}
+
+	if info.IdmType != "" {
+		fmt.Fprintln(ngsi.StdWriter, "IdmType "+info.IdmType)
+	}
+	if info.IdmHost != "" {
+		fmt.Fprintln(ngsi.StdWriter, "IdmHost "+info.IdmHost)
+	}
+	if info.Username != "" {
+		fmt.Fprintln(ngsi.StdWriter, "Username "+info.Username)
+	}
+	if info.Password != "" {
+		fmt.Fprintln(ngsi.StdWriter, "Password "+info.Password)
+	}
+	if info.ClientID != "" {
+		fmt.Fprintln(ngsi.StdWriter, "ClientID "+info.ClientID)
+	}
+	if info.ClientSecret != "" {
+		fmt.Fprintln(ngsi.StdWriter, "ClientSecret "+info.ClientSecret)
+	}
+
+	if info.XAuthToken != "" {
+		fmt.Fprintln(ngsi.StdWriter, "XAuthToken "+info.XAuthToken)
+	}
+	if info.Token != "" {
+		fmt.Fprintln(ngsi.StdWriter, "Token "+info.Token)
+	}
+
+	if info.APIPath != "" {
+		fmt.Fprintln(ngsi.StdWriter, "APIPath "+info.APIPath)
+	}
 }
